# app/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateIssueRequest": {
        "properties": {
          "appliance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliance"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "model_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Number"
          },
          "symptom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symptom"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "user-default",
            "title": "User Id"
          }
        },
        "title": "CreateIssueRequest",
        "type": "object"
      },
      "CreateIssueResponse": {
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "CreateIssueResponse",
        "type": "object"
      },
      "DeleteIssueResponse": {
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "case_id",
          "deleted"
        ],
        "title": "DeleteIssueResponse",
        "type": "object"
      },
...
      "EscalateResponse": {
        "properties": {
          "drafted_email": {
            "title": "Drafted Email",
            "type": "string"
          },
          "escalation_steps": {
            "items": {
              "$ref": "#/components/schemas/EscalationStep"
            },
            "title": "Escalation Steps",
            "type": "array"
          },
          "inspection_guide": {
            "items": {
              "$ref": "#/components/schemas/InspectionShot"
            },
            "title": "Inspection Guide",
            "type": "array"
          },
          "packet": {
            "$ref": "#/components/schemas/Packet"
          }
        },
        "required": [
          "drafted_email",
          "inspection_guide",
          "packet"
        ],
        "title": "EscalateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueDetail": {
        "properties": {
          "appliance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliance"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "diagnosis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diagnosis"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "escalation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Escalation"
              },
              {
                "type": "null"
              }
            ]
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaRef"
            },
            "title": "Media",
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "title": "Messages",
            "type": "array"
          },
          "model_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Number"
          },
          "next_step": {
            "default": "",
            "title": "Next Step",
            "type": "string"
          },
          "status": {
            "enum": [
              "intake",
              "diagnosing",
              "awaiting_user",
              "escalated",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "title": "Steps",
            "type": "array"
          },
          "symptom": {
            "default": "",
            "title": "Symptom",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "IssueDetail",
        "type": "object"
      },
      "IssueSummary": {
        "properties": {
          "appliance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliance"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "model_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Number"
          },
          "next_step": {
            "default": "",
            "title": "Next Step",
            "type": "string"
          },
          "status": {
            "enum": [
              "intake",
              "diagnosing",
              "awaiting_user",
              "escalated",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "symptom": {
            "default": "",
            "title": "Symptom",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "title",
          "status",
          "updated_at"
        ],
        "title": "IssueSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stub REST and SSE contract for the HomeRescue client.",
    "title": "HomeRescue API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/issues": {
      "get": {
        "operationId": "list_issues_api_issues_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "open",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueSummary"
                  },
                  "title": "Response List Issues Api Issues Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Issues"
      },
      "post": {
        "operationId": "create_issue_api_issues_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIssueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Issue"
      }
    },
    "/api/issues/{case_id}": {
      "delete": {
        "operationId": "delete_issue_api_issues__case_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteIssueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Issue"
      },
      "get": {
        "operationId": "get_issue_api_issues__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Issue"
      },
      "post": {
        "operationId": "update_issue_api_issues__case_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Issue"
      }
    },
    "/api/issues/{case_id}/escalate": {
      "post": {
        "operationId": "escalate_issue_api_issues__case_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Escalate Issue"
      }
    },
    "/api/issues/{case_id}/media": {
      "post": {
        "operationId": "upload_media_api_issues__case_id__media_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_api_issues__case_id__media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Media"
      }
    },
    "/api/issues/{case_id}/media/{ref}": {
      "get": {
        "description": "Serve a previously uploaded media file so the client can render it inline in chat.",
        "operationId": "get_media_api_issues__case_id__media__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Media"
      }
    },
    "/api/issues/{case_id}/message": {
      "post": {
        "operationId": "message_issue_api_issues__case_id__message_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message Issue"
      }
    },
    "/api/issues/{case_id}/plate": {
      "post": {
        "operationId": "read_plate_api_issues__case_id__plate_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Plate"
      }
    },
    "/api/issues/{case_id}/resolve": {
      "post": {
        "operationId": "resolve_issue_api_issues__case_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Issue"
      }
    },
    "/api/issues/{case_id}/start": {
      "post": {
        "description": "Auto-kickoff: run ONE agent turn so Gemini produces the first fix for a freshly created\ncase, instead of waiting for the user to type. One-shot and idempotent: only an `intake`\ncase with a symptom is eligible, and the turn moves it out of `intake` so it never re-fires.",
        "operationId": "start_issue_api_issues__case_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Issue"
      }
    }
  }
}